+++ lib/client.js
@@ -1528,7 +1528,7 @@ function write (client, request) {
     if (!expectsPayload) {
       socket[kReset] = true
     }
-  } else if (util.isBlob(body)) {
+  } else if (util.isBlobLike(body)) {
     writeBlob({ client, request, socket, contentLength, header, expectsPayload })
   } else if (util.isStream(body)) {
     writeStream({ client, request, socket, contentLength, header, expectsPayload })
