+++ lib/core/util.js
@@ -18,6 +18,17 @@ function isBlob (obj) {
   return obj && Blob && obj instanceof Blob
 }
 
+// based on https://github.com/node-fetch/fetch-blob/blob/8ab587d34080de94140b54f07168451e7d0b655e/index.js#L229-L241 (MIT License)
+function isBlobLike (object) {
+  return object instanceof Blob || (
+    object &&
+    typeof object === 'object' &&
+    (typeof object.stream === 'function' ||
+      typeof object.arrayBuffer === 'function') &&
+    /^(Blob|File)$/.test(object[Symbol.toStringTag])
+  )
+}
+
 function parseURL (url) {
   if (typeof url === 'string') {
     url = new URL(url)
@@ -266,6 +277,7 @@ module.exports = {
   nop,
   isDisturbed,
   isAborted,
+  isBlobLike,
   parseOrigin,
   parseURL,
   getServerName,
