+++ lib/fetch/util.js
@@ -2,9 +2,9 @@
 
 const { redirectStatus } = require('./constants')
 const { performance } = require('perf_hooks')
-const { Blob } = require('buffer')
 const { File } = require('./file')
 const nodeUtil = require('util')
+const { isBlobLike } = require('../core/util')
 
 let ReadableStream
 
@@ -68,17 +68,6 @@ function requestBadPort (request) {
 
   // 3. Return allowed.
   return 'allowed'
-}
-
-// based on https://github.com/node-fetch/fetch-blob/blob/8ab587d34080de94140b54f07168451e7d0b655e/index.js#L229-L241 (MIT License)
-function isBlobLike (object) {
-  return object instanceof Blob || (
-    object &&
-    typeof object === 'object' &&
-    (typeof object.stream === 'function' ||
-      typeof object.arrayBuffer === 'function') &&
-    /^(Blob|File)$/.test(object[Symbol.toStringTag])
-  )
 }
 
 function isFileLike (object) {
